removes deprecated ingester component
