+++ main.go
@@ -60,7 +60,6 @@ func main() {
 
 	vulcan.AddCommand(cmd.Cacher())
 	vulcan.AddCommand(cmd.Indexer())
-	vulcan.AddCommand(cmd.Ingester())
 	vulcan.AddCommand(cmd.Querier())
 	vulcan.AddCommand(cmd.Forwarder())
 	vulcan.AddCommand(cmd.Downsampler())
